#1955 Removed OpenEntityMasterHelpAction and related classes. Removed master that wraps EntityMasterHelp. Adjusted prodicers and companion objects.
