+++ platform-pojo-bl/src/main/java/ua/com/fielden/platform/domain/PlatformDomainTypes.java
@@ -22,12 +22,11 @@ import ua.com.fielden.platform.entity.AbstractEntity;
 import ua.com.fielden.platform.entity.EntityDeleteAction;
 import ua.com.fielden.platform.entity.EntityEditAction;
 import ua.com.fielden.platform.entity.EntityExportAction;
-import ua.com.fielden.platform.entity.UserDefinableHelp;
 import ua.com.fielden.platform.entity.EntityNewAction;
-import ua.com.fielden.platform.entity.OpenEntityMasterHelpAction;
 import ua.com.fielden.platform.entity.SecurityMatrixInsertionPoint;
 import ua.com.fielden.platform.entity.SecurityMatrixSaveAction;
 import ua.com.fielden.platform.entity.SecurityTokenTreeNodeEntity;
+import ua.com.fielden.platform.entity.UserDefinableHelp;
 import ua.com.fielden.platform.entity.functional.centre.CentreContextHolder;
 import ua.com.fielden.platform.entity.functional.centre.SavingInfoHolder;
 import ua.com.fielden.platform.entity.functional.master.AcknowledgeWarnings;
@@ -174,7 +173,6 @@ public class PlatformDomainTypes {
         types.add(UserMaster_OpenMain_MenuItem.class);
         types.add(UserMaster_OpenUserAndRoleAssociation_MenuItem.class);
         types.add(UserDefinableHelp.class);
-        types.add(OpenEntityMasterHelpAction.class);
 
         typesDependentOnWebUI.add(EntityExportAction.class);
         typesDependentOnWebUI.add(CentreConfigUpdater.class);
