--- platform-pojo-bl/src/main/java/ua/com/fielden/platform/entity/OpenEntityMasterHelpActionCo.java
@@ -1,13 +1,0 @@
-package ua.com.fielden.platform.entity;
-
-import ua.com.fielden.platform.dao.IEntityDao;
-
-/**
- * Companion object for entity {@link OpenEntityMasterHelpAction}.
- *
- * @author TG Team
- *
- */
-public interface OpenEntityMasterHelpActionCo extends IEntityDao<OpenEntityMasterHelpAction> {
-
-}
