+++ platform-pojo-bl/src/test/java/ua/com/fielden/platform/security/provider/SecurityTokenProviderAndNodeConstructionTest.java
@@ -61,7 +61,7 @@ public class SecurityTokenProviderAndNodeConstructionTest {
         assertEquals("Incorrect number of top security tokens.", 35, topNodes.size());
 
         // skip attachment related security tokens before getting iterator nodesWithSkippedAttachmentTokens
-        final Iterator<SecurityTokenNode> superIter = topNodes.stream().skip(19).collect(toList()).iterator();
+        final Iterator<SecurityTokenNode> superIter = topNodes.stream().skip(18).collect(toList()).iterator();
 
         final SecurityTokenNode top1 = superIter.next();
         assertEquals("Incorrect first top token.", Top1LevelSecurityToken.class, top1.getToken());
